+++ packages/terra-abstract-modal/src/AbstractModal.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { Portal } from 'react-portal';
 import KeyCode from 'keycode-js';
 import 'mutationobserver-shim';
+import './_contains-polyfill';
 import './_matches-polyfill';
 import 'wicg-inert';
 import ModalContent from './_ModalContent';
@@ -112,7 +113,7 @@ class AbstractModal extends React.Component {
     this.setState({ modalTrigger: document.activeElement });
 
     if (mainDocumentElement) {
-      const inert = +mainDocumentElement.dataset.overlayCount;
+      const inert = +mainDocumentElement.getAttribute('data-overlay-count');
 
       if (!mainDocumentElement.hasAttribute('data-overlay-count')) {
         mainDocumentElement.setAttribute('data-overlay-count', '1');
@@ -135,7 +136,7 @@ class AbstractModal extends React.Component {
     const mainDocumentElement = document.querySelector(this.props.rootSelector);
 
     if (mainDocumentElement) {
-      const inert = +mainDocumentElement.dataset.overlayCount;
+      const inert = +mainDocumentElement.getAttribute('data-overlay-count');
 
       if (inert === 1) {
         mainDocumentElement.removeAttribute('data-overlay-count');