+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added Node.contains polyfill to better support IE 10 / IE 11
+
+### Fixed
+* Fixed issue in IE 10 when trying to pull value from data attribute
 
 3.6.0 - (July 16, 2019)
 ------------------